Functions protected and revised

+++ ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: jalfonso <jalfonso@student.42abudhabi.ae>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/15 18:30:52 by jalfonso          #+#    #+#             */
-/*   Updated: 2022/08/14 19:21:55 by jalfonso         ###   ########.fr       */
+/*   Updated: 2022/08/15 17:27:46 by jalfonso         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 int	ft_atoi(const char *nptr)
 {
-	long		result;
-	long		sign;
+	int			result;
+	int			sign;
 	size_t		i;
 
 	i = 0;
@@ -47,6 +47,9 @@ int	ft_atoi(const char *nptr)
 //
 // 	printf("ft_atoi: %d\n", ft_atoi("0"));
 // 	printf("ft_atoi: %d\n\n", atoi("0"));
+//
+// 	printf("ft_atoi: %d\n", ft_atoi("2147483647"));
+// 	printf("ft_atoi: %d\n\n", atoi("2147483647"));
 //
 // 	printf("ft_atoi: %d\n", ft_atoi("2147483647"));
 // 	printf("ft_atoi: %d\n\n", atoi("2147483647"));
